+++ common/src/main/java/io/github/mortuusars/thief/Thief.java
@@ -153,8 +153,16 @@ public class Thief {
         }
 
         public static class EntityTypes {
+            public static final TagKey<EntityType<?>> KILLING_PROTECTED_LIGHT =
+                    TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("killing_protected/light"));
+            public static final TagKey<EntityType<?>> KILLING_PROTECTED_MEDIUM =
+                    TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("killing_protected/medium"));
+            public static final TagKey<EntityType<?>> KILLING_PROTECTED_HEAVY =
+                    TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("killing_protected/heavy"));
+
             public static final TagKey<EntityType<?>> WITNESSES =
                     TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("witnesses"));
+
             public static final TagKey<EntityType<?>> GUARDS =
                     TagKey.create(net.minecraft.core.registries.Registries.ENTITY_TYPE, resource("guards"));
         }