+++ fabric/src/main/java/io/github/mortuusars/thief/fabric/ThiefFabric.java
@@ -4,13 +4,18 @@ import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry;
 import io.github.mortuusars.thief.Config;
 import io.github.mortuusars.thief.Thief;
 import io.github.mortuusars.thief.event.CommonEvents;
+import io.github.mortuusars.thief.event.ServerEvents;
 import io.github.mortuusars.thief.network.fabric.PacketsImpl;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
+import net.fabricmc.fabric.api.entity.event.v1.ServerLivingEntityEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.event.player.UseEntityCallback;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraftforge.fml.config.ModConfig;
 import org.jetbrains.annotations.Nullable;
 
@@ -40,6 +45,12 @@ public class ThiefFabric implements ModInitializer {
         UseEntityCallback.EVENT.register((player, world, hand, entity, hitResult) -> {
             if (hitResult != null) return InteractionResult.PASS;
             return CommonEvents.onEntityInteracted(player, hand, entity);
+        });
+
+        ServerLivingEntityEvents.AFTER_DEATH.register((LivingEntity entity, DamageSource damageSource) -> {
+            if (damageSource.getEntity() instanceof ServerPlayer player) {
+                ServerEvents.onEntityKilled(player, entity, damageSource);
+            }
         });
 
         PacketsImpl.registerC2SPackets();